factor slider instrument out, too
